requirements.txt
# --- requirements.txt ---
kafka-python
textblob
elasticsearch